+++ EnumAxis.ts
@@ -1,0 +1,7 @@
+export { Axis };
+
+enum Axis {
+  X,
+  Y,
+  Z,
+}
